Implement "it runs" CI checks
As a part of the merge process, we should make sure that incoming
changes don't break our ability to run tests.  To do this, setup a
simple minikube cluster and attempt to run tests against an empty
cluster.  If the setup resources of a test gets created (i.e. secret
data), and if the service binding sees no change in its status field,
consider it a successful run.

Signed-off-by: Andy Sadler <ansadler@redhat.com>