Push generic app image to ghcr.io on merge
During CI builds, we build and deploy the generic app image directly
into minikube's container runtime.  If that was successful, and that PR
was merged, build and deploy that image to
ghcr.io/servicebindings/generic-app-image.

Container tags are determined by the docker metadata action in this
pipeline.  For reference on how tags are determined, see its
documentation at
https://github.com/marketplace/actions/docker-metadata-action#tags-input.

Signed-off-by: Andy Sadler <ansadler@redhat.com>