lint: don't fork unnecessarily
When we're linting, we call pip to install dependencies.  However, we do
that in every hack/*.sh file, since they could be run standalone. This
means that every time we run a different linter, we call pip and ask it
to make sure our dependencies are installed, which is slow.

This patch changes linting behavior to do pip install once at the
beginning of linting.  However, once that's happened, we can execute
them all in sequence without touching pip.

To make things simpler, this patch also changes the execution behavior
to no longer fork() the bash interpreter for each linter.  This lets us
assume that some setup has already happened, so we only run
prepare_env.sh once

Signed-off-by: Andy Sadler <ansadler@redhat.com>