allow namespace creation to be fallible
The check between `namespace.is_present()` and `namespace.create()`
could have a race condition when running tests in parallel.  Each runner
would see that the namespace doesn't exist, and each would try to create
it.  The correct behavior in this case should be to only create the test
namespace once.

To correct this behavior, allow namespace creation to be fallible.  If
it fails, retry it using a polling strategy.

Signed-off-by: Andy Sadler <ansadler@redhat.com>